// apps/homa-and-mukto/src/common/types.ts
import { PoolClient } from 'pg';

type Record =
{
  id: string;
  created_at: string;
  updated_at: string;
};

export type New<T> = Omit<T, 'id' | 'created_at' | 'updated_at'>

export type Address = Record &
{
  latitude: number;
  longitude: number;
  meta?: {
    address_components: AddressComponent[];
    formatted_address: string;
  };
  user_id: string;
};

export type Credentials =
{
  email: string;
  password: string;
};

export type Registration = Omit<User, 'address.user_id'> & Credentials;

export type User = Record &
{
  name: string;
  image: string;
  contacts: Contact[];
  groups: Group[];
  address?: Address;
};

export type FullUser = User & Credentials &
{
  admin: boolean;
  status: 'verify' | 'review' | 'active';
  pending?:
  {
    name: string;
    image: string;
  };
};

export type Contact =
{
  type: string;
  value: string;
};

export type Group =
{
  type: string;
  location: string;
  month: number;
  year: number;
};

export type Token =
{
  access_token: string;
  access_token_expires_at?: string;
  refresh_token?: string;
  refresh_token_expires_at?: string;
  user_id: string;
  user?: FullUser;
};

export type TokenCamelCase =
{
  accessToken: string;
  accessTokenExpiresAt?: string;
  refreshToken?: string;
  refreshTokenExpiresAt?: string;
  user: FullUser;
};

// API

export type Context =
{
  baseUrl: string;
  client: PoolClient;
  user?: FullUser;
};

// Google Cloud

export type GeocodeResponse =
{
  results:
  {
    address_components: AddressComponent[];
    formatted_address: string;
    geometry:
    {
      bounds: LatLngBounds;
      location: LatLngShort;
      location_type: 'APPROXIMATE' | 'GEOMETRIC_CENTER' | 'RANGE_INTERPOLATED' | 'ROOFTOP';
      viewport: LatLngBounds;
    };
    partial_match: boolean;
    place_id: string;
    plus_code: string;
    types: PlaceType[];
  }[];
  status: 'INVALID_REQUEST' | 'OK' | 'OVER_DAILY_LIMIT' | 'OVER_QUERY_LIMIT' | 'REQUEST_DENIED' | 'UNKNOWN_ERROR' | 'ZERO_RESULTS';
};

export type AddressComponent =
{
  long_name: string;
  short_name: string;
  types: PlaceType[];
};

export type LatLngBounds =
{
  northeast: LatLngShort;
  southwest: LatLngShort;
};

export type LatLngShort =
{
  lat: number;
  lng: number;
};

export type PlacesAutocompleteResponse =
{
  suggestions:
  {
    placePrediction:
    {
      place: string;
      placeId: string;
      text: AutocompleteText;
      structuredFormat:
      {
        mainText: AutocompleteText;
        secondaryText: AutocompleteText;
      };
      types: PlaceType[];
    };
  }[];
};

export type AutocompleteText =
{
  text: string;
  matches?:
  {
    endOffset: number;
  };
};

// There are many more place types, but we only care about these...
export type PlaceType =
  'administrative_area_level_1' |
  'administrative_area_level_2' |
  'administrative_area_level_3' |
  'administrative_area_level_4' |
  'administrative_area_level_5' |
  'administrative_area_level_6' |
  'administrative_area_level_7' |
  'country' |
  'locality';